PAssing to TypeScript WIP
