+++ src/QuizManager.ts
