+++ src/Timer.ts
@@ -1,11 +1,16 @@
-'use strict';
-
 class Timer {
+
+	id: number;
+	callback: Function;
+	delay: number;
+	remaining: number;
+	start: Date;
+
 	constructor() {
 		this.id = null;
 	}
 
-	reset(callback, delay) {
+	reset(callback: Function, delay: number) {
 		if (this.id) {
 			clearTimeout(this.id);
 			this.id = null;
@@ -20,7 +25,7 @@ class Timer {
 		if (this.id) {
 			clearTimeout(this.id);
 			this.id = null;
-			this.remaining = new Date() - this.start;
+			this.remaining = (new Date() as any as number) - (this.start as any as number);
 		}
 	}
 
@@ -32,4 +37,4 @@ class Timer {
 	}
 }
 
-module.exports = Timer;
+export default Timer;
