--- token-example.js
@@ -1,1 +1,0 @@
-module.exports = { token: 'YOUR_BOT_TOKEN_HERE' };
