+++ token-example.json
@@ -1,0 +1,3 @@
+{
+  "token": "YOUR_BOT_TOKEN_HERE"
+}
